catalog + filtration

+++ app/src/main/java/com/example/arsenalmobile/Navigation/NavGraph.kt
@@ -99,7 +99,7 @@ fun NavGraph(
             CatalogBlasters(userController, blasterController, auth, mainActivity, user!!, navHostController)
         }
         composable("${Routes.InfoBlaster.route}/{id}") { navBackStack ->
-            val blasterId = navBackStack.arguments?.getString("id")?.toInt() ?: 0
+            val blasterId = navBackStack.arguments?.getString("id")?.toLong() ?: 0
             InfoBlaster(blasterId = blasterId, blasterController, navHostController, userController, user!!, mainActivity)
         }
 //        composable("${Routes.Material.route}/{id}") { navBackStack ->
